+++ front/src/app/modules/log-dialog/open-log-dialog.service.ts
@@ -1,6 +1,7 @@
 import { MatDialog } from "@angular/material/dialog";
 import { LogDialogComponent } from "./log-dialog/log-dialog.component";
 import { Injectable } from "@angular/core";
+import { IServiceLogs } from "../shell/interfaces/container";
 
 @Injectable()
 export class OpenLogDialogService {
@@ -9,7 +10,7 @@ export class OpenLogDialogService {
     private readonly dialog: MatDialog,
   ) { }
 
-  public openDialog(logs: string) {
+  public openDialog(logs: IServiceLogs) {
     this.dialog.open(
       LogDialogComponent, {
         data: logs,